refactor: 상점 관련 조회서비스 클래스명 수정
