+++ src/main/java/com/example/backend/domain/store/controller/StoreQueryController.java
@@ -8,7 +8,7 @@ import com.example.backend.domain.item.dto.ItemResponse;
 import com.example.backend.domain.item.service.ItemService;
 import com.example.backend.domain.auth.security.CustomUserDetails;
 import com.example.backend.domain.store.dto.StoreResponse;
-import com.example.backend.domain.store.service.StoreService;
+import com.example.backend.domain.store.service.query.StoreQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import java.util.List;
@@ -27,7 +27,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/stores")
 @RequiredArgsConstructor
 public class StoreQueryController {
-    private final StoreService storeService;
+    private final StoreQueryService storeService;
     private final ItemService itemService;
     private final EventService eventService;
