+++ src/main/java/com/example/backend/domain/store/service/query/StoreQueryService.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.store.service;
+package com.example.backend.domain.store.service.query;
 
 import com.example.backend.domain.store.dto.StoreResponse;
 import com.example.backend.domain.store.entity.Store;
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class StoreService {
+public class StoreQueryService {
     private final StoreRepository storeRepository;
 
     public StoreResponse findById(Long storeId) {
